feat: relate users and tasks

+++ nest-exp/src/tasks/tasks.service.ts
@@ -1,12 +1,13 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DeleteResult } from 'typeorm';
 
 import { TaskStatus } from './taskStatus.enum';
 import { CreateTaskDTO } from './dto/createTask.dto';
 import { GetTasksFilterDto } from './dto/getTasksFilter.dto';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
-import { DeleteResult } from 'typeorm';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,15 +15,18 @@ export class TasksService {
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository) { }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user);
   }
-  async createTask(createTaskDto: CreateTaskDTO): Promise<Task> {
-    return await this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDTO,
+    user: User,
+  ): Promise<Task> {
+    return await this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) {
       throw new NotFoundException(`Task with id: ${id} not found`);
     }
@@ -30,12 +34,12 @@ export class TasksService {
     return found;
   }
 
-  async deleteTaskById(id: number): Promise<DeleteResult> {
-    return await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<DeleteResult> {
+    return await this.taskRepository.delete({ id, userId: user.id });
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
